+++ src/Symfony/Bundle/Resources/config/jsonapi.xml
@@ -92,6 +92,11 @@
         <service id="api_platform.jsonapi.listener.request.transform_filtering_parameters" class="ApiPlatform\Symfony\EventListener\JsonApi\TransformFilteringParametersListener">
             <tag name="kernel.event_listener" event="kernel.request" method="onKernelRequest" priority="5" />
         </service>
+
+        <service id="api_platform.json_api.state_provider.default_error" class="ApiPlatform\JsonApi\State\DefaultErrorProvider" >
+            <tag name="api_platform.state_provider" key="ApiPlatform\JsonApi\State\DefaultErrorProvider" />
+            <tag name="api_platform.state_provider" key="api_platform.json_api.state_provider.default_error" />
+        </service>
     </services>
 
 </container>
